faucet: add a zombie channel sweeper
This commit adds a new defensive feature to the faucet: a zombie
channel sweeper. A channel is considered a zombie if the peer we have a
channel open with hasn’t been online in 48 hours, then all the channels
we have open with the peer are to be force closed. We now run this
check every hour in order to cleanse the faucet of any inactive
channels and also prune stale channels from the channel graph.