faucet: omit default port from header on home page
This commit modifies the header within the home page of the faucet to
omit the port from the top of the connection string. A recent change in
lnd switched the default p2p port and also modified the `lncli connect`
command to assume the default port if one isn't specified.